delete and rename few files
